For this commit, to preserve consistency in the names changed the clas segvis_blockList into segvis_block_list. And changed that in every code and documentation.
Changed the method plot.profile of ProfileMatrixList into
plot_profiles of segvis_block_list, it works similarly, but it is
using the summarize method instead of meanProfile. Therefore, it is
possible to use any function which uses a vector as an argument
instead of only mean, trimmed means and median.

Also, created the plot_data method, which returns the same data.table
that plot_data uses to create the ggplot object that returns. This
way, the user could make more organized plots by adding the our
conditions and stuff.

Changed all the documentation accordingly.